refactor: reorganize secret injection strategy types and improve imports
