+++ packages/core/src/types.ts
@@ -8,6 +8,24 @@ export type AnyFunction = FunctionLike<unknown[], unknown>;
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type AnyClass = { new(...args: any[]): any };
 
+export interface BaseSecretInjectionStrategy {
+  /**
+   * Override the default target path for the secret injection.
+   * 
+   * Moreover, each provider has a default target path for the secret injection.
+   * By using BaseProvider.getTargetPath() 
+   */
+  targetPath?: string;
+}
+
+export type SecretInjectionStrategy =
+  | { kind: 'env'; containerIndex?: number } & BaseSecretInjectionStrategy
+  | { kind: 'volume'; mountPath: string; containerIndex?: number } & BaseSecretInjectionStrategy
+  | { kind: 'annotation' } & BaseSecretInjectionStrategy
+  | { kind: 'imagePullSecret' } & BaseSecretInjectionStrategy
+  | { kind: 'envFrom'; containerIndex?: number } & BaseSecretInjectionStrategy
+  | { kind: 'plugin'; action?: string; args?: unknown[]; [key: string]: unknown; };
+
 /**
  * Accept any type of key, including string, number, or symbol, Like `keyof any`.
  * This is useful for generic programming where the key type is not known in advance.
